Geoclustering_EPC/Dash_app/requirements.txt
dash==2.18.2
dash-core-components==2.0.0
dash-echarts==0.0.12.9
dash-html-components==2.0.0
dash-iconify==0.1.2
dash_mantine_components==0.15.1
fastapi==0.115.6
Flask==3.0.3
Flask-Login==0.6.3
numpy==2.1.3
pandas==2.2.3
plotly==5.24.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
uvicorn==0.34.0
scipy==1.15.1
seaborn==0.13.2
urllib3==2.3.0
Werkzeug==3.0.6
workalendar==17.0.0
gunicorn==23.0.0
dash-leaflet==1.0.15
dash_mantine_react_table==0.0.1
xgboost ==2.1.4
lightgbm == 4.6.0
joblib == 1.4.2
tensorflow == 2.19.0
optuna ==4.3.0

